VERSION: PRINT DATA 1

+++ robot/sensors.cpp
@@ -4,4 +4,4 @@
 // Capteurs
 const int MesureMaxi = 50; // Distance maxi a mesurer 
 const int MesureMini = 5; // Distance mini a mesurer 
-long measuredLenght[NB_OF_SENSORS];
+double measuredLenght[NB_OF_SENSORS];
